put GUI code in its own file
